Found bug in whisper.merge(), which is used by script whisper-merge.py. The issue is that fromTime was not reset to current time when it started to look merge changes for the latest Archive.  As such it would complain that fromTime was in the future.
Changed it so that 'now' is set to time.time() and uses same 'now' for all
Archives in the whisper file (vs. potentially changing for each
iteration).  Also moved the setting of fromTime to inside the iterator for
archives, so it is looking back in time instead of into the future.