Adds new absmax aggregation method
absmax is a new aggregation method which returns the largest
absolute value when aggregating to a lower precision archive.

Useful for data such as time offsets, where you care about retaining
the value furthest from zero when aggregating but would prefer to
preserve whether the offset was positive or negative.