Add AssertRaisesException to test_whisper
Unlike unittest.TestCase.assertRaises, this context manager will
also validate the contents of the raised exception and not just
the type.

There are places in the whisper codebase where an InvalidConfiguration
or CorruptWhisperFile exception is raised with different values. This
ensures we are getting the expected exceptions when we think we are.