Implement configurable fetch strategy: coarse
The idea is to return datapoints from the coarsest archive that has above
'threshold' datapoints in the requested timespan.
If such an archive is not available, it returns the first archive matching
the timespan. (as without this patch)

With appropriate retention periods and thresholds configuration,
using this fetch strategy can yield a major speed increase, since
much less datapoints need to be processed.
(sent to a frontend and/or rendered into a graph)

The caller of fetch() will need to set FETCH_STRATEGY to 'coarse'
and FETCH_STRATEGY_COARSE_THRESHOLD to the desired minimum datapoints.

The default strategy is unchanged and the default threshold is
equivalent to the default fetch strategy.