Backport to 0.9.x enforce closing of files
The original commit:
    f610a6505a76df4f4dca9099c0d9c6824b80cb57

This was cleaned up significantly on the master branch.  This is a
backport that wraps open files using Python's with statement to ensure
that opened files are always closed.

This solves a bug that I encountered where a traceback would occur in
whisper.update_many() leaving the .wsp file unclosed and flock()'d.  The
Python GC would not reclaim the out of scope file descriptor and close the
file before carbon-cache's writer thread needed to update the same .wsp
file again.  Issue #129