whisper.py: numpy alternative for unpack & format
Benchmark per number of points fetched, in seconds:
+---------------+------------+-------------+----------+
|   Points       |  Numpy  |    Struct   |  Ratio   |
+---------------+------------+-------------+----------+
|  1                | 0.00326 | 0.0023     |  -29%   |
|  100            | 0.00353 | 0.00475   |   35%   |
|  10,000       | 0.01675 | 0.18555   | 1008% |
|  1,000,000  | 1.64262 | 29.75022 | 1711% |
+---------------+------------+-------------+----------+

The numpy unpacking and formatting option is faster than the struct option when fetching more than 50 points. This method speeds up exponentially as the number of points increases but is slower of 0.001 second when fetching 1 single point.

I also noticed that whisper does not release memory when used via a daemon service. This leaded to critical memory issues on our side.
After investigating this issue, I attributed it to the struct.unpack automatic caching.
Manual cache clearing after each request partially solved the problem but at the cost of huge CPU usage.

This numpy option fixes this issue, releasing all memory after request with minimal CPU usage.
On top of that, the values are returned as a numpy array allowing easy computation transformation and can be jsonified through .tolist()

We use it at flockpit.com. Thank you for your amazing work on Graphite!