fix three bugs making aggregation impossible
The aggregation feature implemented before wasn't working at all.
In most cases there was a loss of data for the smallest retetion time
and some other ugly bugs for bigger retetention times. Also it could
happen that you lose only a few timestamps but in a regular pattern.
So the old version was shredding your data.
This commit is fixing all this issues an was testet to be working on
over 30.000 whisper files.