whisper-fetch: fix timestr when using --drop
Filtering the values list when --drop is set happens too early for the
non-json output.  When looping over the values we expect to be able to
set the time string based on the number of times through the loop,
adding the step value on each pass.  With empty values stripped, the
time gets further off for each value that is dropped.

Move the drop filtering function to the appropriate places for both json
and plain test output.

Testing on databases with more than a year of data and many empty
values, the time to execute is no different, so we should not be
regressing in any significant way.  Though even if we were, it is better
to be accurate than fast if one or the other must be picked.

Fixes #250, fixes #305.